WIP: implement preview for documents with "next dataset"/"next record" fields
This is work in progress and should NOT be merged!

The goal is to show correct values for all fields in the preview even if
the "next dataset" functionality is used in the document. Currently, all fields display
the values for one datset only (which is incorrect).

This version uses LibreOffice's/OpenOffice's "Data to fields"
functionality to insert values into fields for the preview

This version was used to explore the "Data to fields" functionality.

Currently, a fixed data source is used and not all types of fields are
updated correctly. Special preconditions must be met so that values are
actually inserted in the preview (e.g. a data source with the name "DataSource"
must be registered).