trac#14244: FONT_ZOOM has no effect in WollMuxBar / make menumanager & options-menu configurable
1) Bugfix (FONT_ZOOM has no effect in WollMuxBar):

Removed WollMuxFiles.setLookAndFeel(ConfigThingy conf) as this function doesn't reflect the desired logic for defaulting from wollmuxbar.conf to wollmux.conf.

Instead, retrieve FONT_ZOOM settings from the WollMuxBarConfig-instance and call WollMuxFiles.zoomFonts(config.getFontZoom()) manually from WollMuxBar (this methode needed to get public therfore).

2) New Configuration option ALLOW_USER_CONFIG:

This options allows to disable the usage of the wollmuxbar.conf-file completely. This implies that if deactivated, the Buttons with the ACTIONs "options" and "menuManager" are not shown and accessible in the WollMuxBar. The option is initialized with the default value "true".

3) New Configuration option ALLOW_MENUMANAGER:

The menumanager is now considered an experimental feature and hence deactivated by default. The option ALLOW_MENUMANAGER allows to enable this feature. If the feature is deactivated, WollMuxBar-Buttons with the ACTION "menuManager" are not shown and accessible in the WollMuxBar. Now, the menumanager is only accessible in the following conditions:

- ALLOW_USER_CONFIG=true AND ALLOW_MENUMANAGER=true
- or WollMuxBar is called with parameter "-mm" (so -mm switches on ALLOW_USER_CONFIG and ALLOW_MENUMANAGER implicitely)