trac#22298: Consistent 'key not found' vs. 'key value invalid'
use the same PERSISTENT_DATA_MODE value when the key is not present
in wollmux config vs. when it is present, but its value is invalid

This band-aid prevents the crash described in trac#22298, as
defaulting to 'rdfReadLegacy' tries to modify the template. This is
fatal in combo with 6b85a5bfeb2c87e68f0f5dfaa76b23c5feab1bda