trac#30187 fix loading of Styles
with commit f4ddff7db4781767c44ae116e2e563d579a88641 caching was enabled,
but this breaks the loading of styles as the method
XStyleLoader.loadStylesFromURL(url, props) does not support urls with
"private:stream" and therefore does nothing so the styles were not loaded.

As a temporary solution until libreoffice supports loading of styles via
stream, the caching is disabled for styles.