trac#31585 don't mix simulation results with form model
during mailmerge each dataset is initialized with the values from the document,
which have been set during preview. The values from the datasource are set via
formmodel, which only updates it's values if they change. So if there are two
datasource entries with the same value and the dataset is initialized with an
other value the second entry from the datasource has the old values from the
document.