Allow overriding maven.javadoc.failOnWarnings
The javadoc target of the current WollMux 18.02 build fails for
me due to the failOnWarnings setting. From a bug report[1] I
learned, that a user just can override POM properies with a
commandline value, usally refered to as "User property" in the
documentation. The configuration is considered fixed.

So this drops the configuration value and adds the corresponding
property instead.

[1] https://issues.apache.org/jira/browse/MNG-4979