trac#33889 remove print block not working
If there are several print blocks of the same type (eg copy only), they
can't be deleted by pressing the button again. The problem is the
bookmarks have a number at the end so those not equal the command. So we
now check if the bookmark name starts with the command and if it those
we delete it.