trac#30220 do not remove hidden bookmarks in OOoBasedMailMerge
OOoBasedMailMerge.removeAllBookmarks():

Hidden bookmarks with conditions are required during mail merge as the
condition often depends on the current data record.

Apparently this function was added due to performance issues with a
large number of data records in commit
a1549bf958b59d066564fcdfdf81eed7f2135bb0 - remains to be seen if this is
still a problem.